+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is loosely based on [Keep a Changelog](http://keepachangelog.com/).
 ## [v0.30.1] - In Progress
 ### Added
 - Added ability to load a Cuefile into the current Maestro instead of replacing it entirely.
+- Added ability to resize MaestroDrawingArea.
 
 ### Changed
 - Renamed `Cue Interpreter` to `Cue Previewer`.